MFC r293015:
Merge r293013 from clang380-import branch:

Fix a clang 3.8.0 warning in pflogd.c:

contrib/pf/pflogd/pflogd.c:769:8: error: logical not is only applied to the left hand side of this comparison [-Werror,-Wlogical-not-parentheses]
		       if (!if_exists(interface) == -1) {
					   ^                     ~~

The if_exists() function does not return -1, and even if it did, it
would not be the correct way to check.  Just ditch the == -1 instead.

Obtained from:	OpenBSD's pflogd.c 1.49


git-svn-id: https://svn.freebsd.org/base/stable/10@293110 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f