MFC 293043
Unset the gss kernel state when gssd exits

When gssd exits it leaves the kernel state set by
gssd_syscall().  nfsd sees this and waits endlessly
in an unkillable state for gssd to come back.  If you
had acidentally started gssd then stopped it, then
started nfsd you'd be in a bad way until you either
restarted gssd or rebooted the system.  This change
fixes that by setting the kernel state to "" when
gssd exits.


git-svn-id: https://svn.freebsd.org/base/stable/10@293446 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f