MFC r287964:
Kernel part of reroot support - a way to change rootfs without reboot.

Note that the mountlist manipulations are somewhat fragile, and not very
pretty.  The reason for this is to avoid changing vfs_mountroot(), which
is (obviously) rather mission-critical, but not very well documented,
and thus hard to test properly.  It might be possible to rework it to use
its own simple root mount mechanism instead of vfs_mountroot().

Sponsored by:	The FreeBSD Foundation
Differential Revision:	https://reviews.freebsd.org/D2698


git-svn-id: https://svn.freebsd.org/base/stable/10@293743 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f