Remove an unneeded assignment of the return value.
tdelete() is supposed to return the address of the parent node that has
been deleted. We already keep track of this node in the loop between
lines 94-107. The GO_LEFT()/GO_RIGHT() macros are used later on as well,
so we must make sure not to change it to something else.


git-svn-id: https://svn.freebsd.org/base/head@293879 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f