MFC r291391
sfxge: do not use unnamed union in siena_mc_combo_rom_hdr_t

GCC 4.2.1 used on FreeBSD 8 and 9 branches does not like unnamed
union member in the structure. It is not strictly required in head,
but nice to have to minimize difference with out-of-tree driver.

Sponsored by:   Solarflare Communications, Inc.


git-svn-id: https://svn.freebsd.org/base/stable/10@293917 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f