MFstable/10 r294239:
MFC r293704:

Fix theoretical leak of netconfig(3) resources in svcunix_create(..)

In the event that the getconfig(3) call in svcunix_create is partly successful,
some of the netconfig(3) resources allocated might be leaked if the call returns
NULL as endnetconfig(3) wasn't called explicitly in that case. Ensure that the
resources are fully cleaned up by going to the `done` label, which will call
endnetconfig(3) for us.

Submitted by: Miles Ohlrich <miles.ohlrich@isilon.com>


git-svn-id: https://svn.freebsd.org/base/stable/9@294241 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f