[iwm] fix up the rate control setup code to initialise rates in the order we want to use it.
The rate table was being initialised in low->high, but the link quality
table was being initialised high->low. So, when we did a lookup, we
would get the indexes wrong.

This started by a patch from dragonflybsd which reversed how the ni->in_ridx[]
array is being used; I'd rather it all be consistent.  So, this is consistent.

Inspired by: what I did to iwn(4) a while ago
Inspired by: DragonflyBSD; <imre@vdsz.com>


git-svn-id: https://svn.freebsd.org/base/head@294248 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f