MFC r293856:
Avoid reading pass the end of the source buffer when it is not NUL
terminated.

If this buffer is adjacent to an unmapped page or a version of C with
bounds checked is used this may result in a crash.

PR:		206178
Submitted by:	Alexander Cherepanov <cherepan@mccme.ru>
Requested by:	danfe


git-svn-id: https://svn.freebsd.org/base/stable/8@294538 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f