MFC 286974,291653:
286974:
Remove reference to non-existent kern_openat(9).

291653:
The cdevpriv_dtr_t typedef was not able to be used in a function prototype
like the various d_*_t typedefs since it declared a function pointer rather
than a function.  Add a new d_priv_dtor_t typedef that declares the function
and can be used as a function prototype.  The previous typedef wasn't
useful outside of the cdevpriv implementation, so retire it.

The name d_priv_dtor_t was chosen to be more consistent with cdev methods
since it is commonly used in place of d_close_t even though it is not a
direct pointer in struct cdevsw.


git-svn-id: https://svn.freebsd.org/base/stable/10@294612 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f