MFC r294347
Fix usr.bin.truncate.truncate_test.bad_truncate with ZFS /tmp.

The bad_truncate test sets the uimmutable flag to produce an error in
truncate, but that flag isn't supported by ZFS.  If /tmp is on a ZFS
filesystem, the test will fail.  Change it to use readonly permissions and
an unpriveleged user instead.


git-svn-id: https://svn.freebsd.org/base/stable/10@294850 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f