MFC: r295032
Use '^[>+][^+]' instead of '^[>+]' with grep(1) when filtering the
diff(1) output between two files in "new_only"-mode. Otherwise,
with the default of using unified format a remnant of the header
in the output is the result. This is especially irritating when
the two files differ but the second one is empty, amounting to the
vestige of the header being the only readout.
Reported by: Stefan Haemmerl

Approved by:	re (delphij)


git-svn-id: https://svn.freebsd.org/base/stable/10@295130 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f