o Gather all mbuf(9) allocation functions into kern_mbuf.c, and all   mbuf(9) manipulation functions into uipc_mbuf.c.  This looks like   the initial intent, but had diffused in the last decade.
o Gather all declarations in mbuf.h in one place and sort them.

o Uninline m_clget() and m_cljget().

There are no functional changes in this patch.

The patch comes from a larger version, where all mbuf(9) allocation was
uninlined, which allowed to make mbuf(9) UMA zones private to kern_mbuf.c.
The performance impact of the total uninlining is still unclear, so we
are holding on now with larger version.

Together with:	melifaro, olivier


git-svn-id: https://svn.freebsd.org/base/head@295547 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f